+++ mcp.go
@@ -43,7 +43,11 @@ type Session interface {
 	Send(ctx context.Context, msg JSONRPCMessage) error
 
 	// Messages returns an iterator that yields messages received from the client.
+	// The implementations should exit the iteration if the session is closed or interrupted.
 	Messages() iter.Seq[JSONRPCMessage]
+
+	// Interrupt interrupts the session and terminates all active operations.
+	Interrupt()
 }
 
 // Server interfaces